requirements.txt
requests
Pillow
pytesseract
PyMuPDF
PyPDF2
unidecode
google-generativeai